Delete Match_R Code
Founds some inconsistencies in the assignment of projects, specifically because we altered the preferences of the projects to be all 1, essentially meaning they had no preference the matching algo used the order of students being assigned as the first criterion for ranking. Meaning students being assigned first got a higher likelihood of getting their first preference.